Migrate WinGetSVC pipeline to 1ES (#128720)
* Migrate WinGetSVC pipeline to 1ES

This commit includes all the 1ES pipeline migration changes that are working in winget-pkgs-preprod
manually copied to winget-pkgs repo as some issue in automatic deployment didn't bring this changes
automatically.

This commit migrates the WinGetSVC pipeline to 1ES, including changes to the following files:
- publish-pipeline.yaml
- rebuild-pipeline.yaml
- rebuild-rest-pipeline.yaml
- validation-pipeline.yaml

The changes were manually copied from the winget-pkgs-preprod repo to the winget-pkgs repo due to an issue with automatic deployment.

[How Validated:]
- The changes have been validated against the preprod environment by manually triggering the pipeline and ensuring successful runs.

* Improve consistency of pipeline stage names

This commit updates the pipeline stage names to be more consistent across all pipelines. This change improves the readability of the pipeline code.